Fold broadcast test into connection_expr_function_paths
The file duplicated StereoConst, StereoSink, and mod dsp from its
siblings to hold a single graph and test; move them into
connection_expr_function_paths.rs, which already defines the fixtures.
