+++ oscen-lib/tests/connection_expr_function_paths.rs
@@ -2,7 +2,8 @@
 //! source is a call to a path-qualified pure function, e.g.
 //! `dsp::decode_ms(s.output) -> out`. Regression coverage keeps the bare-ident
 //! call and the turbofish frame constructor parsing after the `Ident -> Path`
-//! change.
+//! change, and locks the intersection with Gap B (type-directed array
+//! broadcast): a path-qualified frame function into a node array.
 #![feature(inherent_associated_types)]
 
 use float_cmp::approx_eq;
@@ -81,6 +82,36 @@ impl SignalProcessor for StereoConst {
     }
 }
 
+/// Latches a `Frame<2>` stream input every sample (frame-broadcast sink).
+#[derive(Debug, Node)]
+pub struct StereoSink {
+    #[input(stream)]
+    pub input: Frame<2>,
+    pub last: Frame<2>,
+}
+
+impl StereoSink {
+    pub fn new() -> Self {
+        Self {
+            input: Frame([0.0; 2]),
+            last: Frame([0.0; 2]),
+        }
+    }
+}
+
+impl Default for StereoSink {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl SignalProcessor for StereoSink {
+    #[inline(always)]
+    fn process(&mut self) {
+        self.last = self.input;
+    }
+}
+
 // path-qualified function: `dsp::decode_ms` applied on a connection.
 graph! {
     name: PathFnGraph;
@@ -126,6 +157,20 @@ graph! {
 
     connections {
         oscen::frame::Frame::<2>(a.output, b.output) -> out;
+    }
+}
+
+// path-qualified frame function broadcast into a Frame<2>-input node array.
+graph! {
+    name: PathFnBroadcastGraph;
+
+    nodes {
+        s = StereoConst::new();
+        sinks = [StereoSink::new(); 3];
+    }
+
+    connections {
+        dsp::decode_ms(s.output) -> sinks.input;
     }
 }
 
@@ -183,6 +228,24 @@ fn qualified_frame_constructor_matches_bare() {
         "expected Frame([0.25, -0.5]), got {:?}",
         graph.out
     );
+}
+
+#[test]
+fn path_qualified_function_broadcasts_into_node_array() {
+    let mut graph = PathFnBroadcastGraph::new();
+    graph.init(RATE);
+    // mid = 0.6, side = 0.1 -> left = 0.5, right = 0.7, latched on every element.
+    graph.s.val = Frame([0.6, 0.1]);
+    graph.process();
+    for i in 0..3 {
+        assert!(
+            approx_eq!(f32, graph.sinks[i].last.0[0], 0.5, epsilon = 1e-6)
+                && approx_eq!(f32, graph.sinks[i].last.0[1], 0.7, epsilon = 1e-6),
+            "sinks[{}].last expected Frame([0.5, 0.7]), got {:?}",
+            i,
+            graph.sinks[i].last
+        );
+    }
 }
 
 #[test]
